requirements.txt
temporalio==1.6.0
redis==5.0.1
boto3==1.34.0
pydantic==2.5.0
click==8.1.7
python-dotenv==1.0.0
